requirements.txt
aiohttp==3.9.1
aiomysql==0.3.2
alembic==1.17.0
altair==5.5.0
anyio==4.11.0
attrs==25.4.0
bcrypt==3.2.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.62.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
joblib==1.3.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
narwhals==2.9.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
pip==25.2
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.24
pydeck==0.9.1
PyJWT==2.8.0
PyMySQL==1.1.0
redis==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
scikit-learn==1.3.2
rpds-py==0.27.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.13.6
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
